Query enroll for cohorts using new graphql endpoint
As we begin to want to use Census for authentication with other services
(we're starting with Full Circle) we want to stop duplicating the data
and manual creation of cohorts all over the place.

Enroll is the source of truth for Cohorts (Enroll calls them "sections")
and other apps that need access to them should fetch them from Enroll
and persist id pointers (instead of making their own copies).

This changeset updates Census to fetch cohorts from Enroll via a Graphql
client. The client is configurable via `API_AUTH_SECRET`, which must
match what Enroll has and `ENROLL_GRAPHQL_ENDPOINT`, which can be
configured to run against an arbitrary endpoint. All cohort management
should happen in Enroll so mutating cohort functionality has been
removed.

Add rake task to dump to disk the enroll graphql schema. Meant to be run
locally and committed. This task must be run and committed when the
Enroll API changes. When we move to a client this work can be moved to
the client libray instead.

Updates the specs to mock out `Cohort.all` to return a stubbed response,
which can be overridden per spec as needed.

Future work:
* Update Enroll to write to census somehow to keep user cohorts / groups
  up to date. This is not critical for now but will be as we want more
  apps to have up to date data about users.
* Update the usage of `graphql-client` to be wrapped in a client gem
  that we can share across many apps. We'll extract it once we've used
  it again and see what patterns we need.

https://trello.com/c/Ov6egTAg/240-have-census-pull-cohorts-from-enroll