Remove cohorts from registration form
the sign up flow after getting invited to census (with enroll access)
has no cohort set (since they are not yet enrolled) but the sing up page
in census allows the user to pick a cohort. We want to remove that
cohort selection since they are going to do that in Enroll.

we also want enrollment to always happen in Enroll so we're removing the
cohort dropdown from the registration for entirely. It is still seen on
the user edit page (which admins can use).

https://trello.com/c/26SqGzIn/322-as-an-invited-user-i-should-not-be-able-to-pick-a-cohort