Scope add / remove roles to be under admin API namespace
Adding or removing a role from a user is a sensitive operation since it
allows role escalation or removal. This API is accessed mainly by admins
logged into the UI but could also be access by admin scoped oauth apps.

This changeset moves the add_roles and remove_roles endpoints to be
under the api/admin namespace, which requires and admin user (via the
browser) or an admin scope (via token).

https://trello.com/c/DkKNuk2r/278-make-apis-accessed-by-ui-require-currentuser-be-an-admin