Use simple mail for invitation resend
When a user is invited via Apply they get the "welcome to turing" email.

When we resend the invitation (via an admin action in census) it's
confusing for the user to get the same message again.

This change updates the resend endpoint to pass a flag so we send the
simple template upon resending.

Follow up from
https://trello.com/c/0JU5dWA8/414-add-name-to-census-invite-email-from-apply

https://trello.com/c/FQ4158sc/417-update-census-invitation-resend-logic-to-not-send-the-invitation-email